PISP under PSD2_SCT flows_Shopper PISP
Case where the SHopper chooses his PISP.
This PISP manages the initiation dialog and generates the transfer initiation message to the Shopper's Bank (which could be the Shopper Account servicing PSP) and gets in return a report (status).
Remains the problem of strong authentication since the PISP has not authentication method available today to prove to the bank that the shopper is really who he pretends to be. Cf. SCAI approach to solve this.
As a reminder, PSD2 says an important thing about the question:
"
(30) The personalised security credentials used for secure customer authentication by the payment service user or by the payment initiation service provider are usually those issued by the account servicing payment service providers. Payment initiation service providers do not necessarily enter into a contractual relationship with the account servicing payment service providers and, regardless of the business model used by the payment initiation service providers, the account servicing payment service providers should make it possible for payment initiation service providers to rely on the authentication procedures provided by the account servicing payments service providers to initiate a specific payment on behalf of the payer."